refactor: Refactor variables controller into a RestController
