+++ packages/cli/src/environments/variables/variables.controller.ee.ts
@@ -1,43 +1,51 @@
-import express from 'express';
-import { Container } from 'typedi';
+import { Container, Service } from 'typedi';
 
 import * as ResponseHelper from '@/ResponseHelper';
-import type { VariablesRequest } from '@/requests';
+import { VariablesRequest } from '@/requests';
+import { Authorized, Delete, Get, Patch, Post, RestController } from '@/decorators';
 import {
+	VariablesService,
 	VariablesLicenseError,
-	EEVariablesService,
 	VariablesValidationError,
 } from './variables.service.ee';
 import { isVariablesEnabled } from './enviromentHelpers';
 import { Logger } from '@/Logger';
+import type { RequestHandler } from 'express';
 
-export const EEVariablesController = express.Router();
+const variablesLicensedMiddleware: RequestHandler = (req, res, next) => {
+	if (isVariablesEnabled()) {
+		next();
+	} else {
+		res.status(403).json({ status: 'error', message: 'Unauthorized' });
+	}
+};
 
-EEVariablesController.use((req, res, next) => {
-	if (!isVariablesEnabled()) {
-		next('router');
-		return;
+@Service()
+@Authorized()
+@RestController('/variables')
+export class VariablesController {
+	constructor(
+		private variablesService: VariablesService,
+		private logger: Logger,
+	) {}
+
+	@Get('/')
+	async getVariables() {
+		return Container.get(VariablesService).getAllCached();
 	}
 
-	next();
-});
-
-EEVariablesController.post(
-	'/',
-	ResponseHelper.send(async (req: VariablesRequest.Create) => {
+	@Post('/', { middlewares: [variablesLicensedMiddleware] })
+	async createVariable(req: VariablesRequest.Create) {
 		if (req.user.globalRole.name !== 'owner') {
-			Container.get(Logger).info(
-				'Attempt to update a variable blocked due to lack of permissions',
-				{
-					userId: req.user.id,
-				},
-			);
-			throw new ResponseHelper.AuthError('Unauthorized');
+			this.logger.info('Attempt to update a variable blocked due to lack of permissions', {
+				userId: req.user.id,
+			});
+			throw new ResponseHelper.UnauthorizedError('Unauthorized');
 		}
 		const variable = req.body;
 		delete variable.id;
 		try {
-			return await Container.get(EEVariablesService).create(variable);
+			return await Container.get(VariablesService).create(variable);
 		} catch (error) {
 			if (error instanceof VariablesLicenseError) {
 				throw new ResponseHelper.BadRequestError(error.message);
@@ -46,27 +54,32 @@ EEVariablesController.post(
 			}
 			throw error;
 		}
-	}),
-);
+	}
 
-EEVariablesController.patch(
-	'/:id(\\w+)',
-	ResponseHelper.send(async (req: VariablesRequest.Update) => {
+	@Get('/:id(\\w+)')
+	async getVariable(req: VariablesRequest.Get) {
+		const id = req.params.id;
+		const variable = await Container.get(VariablesService).getCached(id);
+		if (variable === null) {
+			throw new ResponseHelper.NotFoundError(`Variable with id ${req.params.id} not found`);
+		}
+		return variable;
+	}
+
+	@Patch('/:id(\\w+)', { middlewares: [variablesLicensedMiddleware] })
+	async updateVariable(req: VariablesRequest.Update) {
 		const id = req.params.id;
 		if (req.user.globalRole.name !== 'owner') {
-			Container.get(Logger).info(
-				'Attempt to update a variable blocked due to lack of permissions',
-				{
-					id,
-					userId: req.user.id,
-				},
-			);
-			throw new ResponseHelper.AuthError('Unauthorized');
+			this.logger.info('Attempt to update a variable blocked due to lack of permissions', {
+				id,
+				userId: req.user.id,
+			});
+			throw new ResponseHelper.UnauthorizedError('Unauthorized');
 		}
 		const variable = req.body;
 		delete variable.id;
 		try {
-			return await Container.get(EEVariablesService).update(id, variable);
+			return await Container.get(VariablesService).update(id, variable);
 		} catch (error) {
 			if (error instanceof VariablesLicenseError) {
 				throw new ResponseHelper.BadRequestError(error.message);
@@ -75,5 +88,20 @@ EEVariablesController.patch(
 			}
 			throw error;
 		}
-	}),
-);
+	}
+
+	@Delete('/:id(\\w+)')
+	async deleteVariable(req: VariablesRequest.Delete) {
+		const id = req.params.id;
+		if (req.user.globalRole.name !== 'owner') {
+			this.logger.info('Attempt to delete a variable blocked due to lack of permissions', {
+				id,
+				userId: req.user.id,
+			});
+			throw new ResponseHelper.UnauthorizedError('Unauthorized');
+		}
+		await this.variablesService.delete(id);
+
+		return true;
+	}
+}
